refactor: move lstbin_simple to lst_stack subpackage
