+++ hera_cal/lstbin.py
@@ -21,15 +21,15 @@ from . import redcal
 from . import apply_cal
 from .datacontainer import DataContainer
 import logging
-from pyuvdata.uvdata import FastUVH5Meta
-from pyuvdata.utils import get_lst_for_time
-try:
-    profile
-except NameError:
-    def profile(fnc):
-        return fnc
+import warnings
 
 logger = logging.getLogger(__name__)
+
+warnings.warn(
+    "This module is deprecated and will be removed in a future version of hera_cal. ",
+    DeprecationWarning,
+)
+
 
 def baselines_same_across_nights(data_list):
     """
@@ -58,6 +58,7 @@ def baselines_same_across_nights(data_list):
                 baseline_counts[k] = 1
     same_across_nights = np.all([baseline_counts[k] == baseline_counts[bl] for bl in baseline_counts])
     return same_across_nights
+
 
 @profile
 def lst_bin(data_list, lst_list, flags_list=None, nsamples_list=None, dlst=None, begin_lst=None, lst_low=None,
@@ -211,9 +212,9 @@ def lst_bin(data_list, lst_list, flags_list=None, nsamples_list=None, dlst=None,
         # iterate over keys in d
         klist = list(d.keys())
         for j, key in enumerate(klist):
-            if j % max(1, (len(klist)//100)) == 0:
-                logger.info(f"Doing key {key} [{j+1}/{len(klist)}]")
-            
+            if j % max(1, (len(klist) // 100)) == 0:
+                logger.info(f"Doing key {key} [{j + 1}/{len(klist)}]")
+
             # if bl_list is not None, use it to determine conjugation:
             # this is to prevent situations where conjugation of bl in
             # data_list is different from bl in data which can cause
@@ -316,7 +317,6 @@ def lst_bin(data_list, lst_list, flags_list=None, nsamples_list=None, dlst=None,
     # wrap lst_bins if needed
     lst_bins = lst_bins % (2 * np.pi)
 
-
     # return un-averaged data if desired
     if return_no_avg:
         # return all binned data instead of just the bin average
@@ -346,7 +346,7 @@ def lst_bin(data_list, lst_list, flags_list=None, nsamples_list=None, dlst=None,
         bin_count = []
         # iterate over sorted LST grid indices in data[key]
         for j, ind in enumerate(sorted(data[key].keys())):
-            
+
             # make data and flag arrays from lists
             d = np.array(data[key][ind])  # shape = (Ndays x Nfreqs)
             f = np.array(flags[key][ind])
@@ -405,13 +405,13 @@ def lst_bin(data_list, lst_list, flags_list=None, nsamples_list=None, dlst=None,
                 mask = norm > 0
                 real_avg_t = np.full(d.shape[1:], np.nan)
                 imag_avg_t = np.full(d.shape[1:], np.nan)
-                
+
                 sm = np.nansum(d * n, axis=0)
                 sm[mask] /= norm[mask]
 
                 real_avg_t[mask] = sm.real[mask]
-                imag_avg_t[norm>0] = sm.imag[mask]
-                
+                imag_avg_t[norm > 0] = sm.imag[mask]
+
                 # add back nans as np.nansum sums nan slices to 0
                 flagged_f = np.logical_not(isfinite).all(axis=0)
                 real_avg_t[flagged_f] = np.nan
@@ -763,7 +763,7 @@ def lst_bin_files(data_files, input_cals=None, dlst=None, verbose=True, ntimes_p
     if Nblgroups is None:
         if Nbls_to_load in [None, 'None', 'none']:
             Nbls_to_load = len(bl_nightly_dicts) + 1
-        
+
         Nblgroups = len(bl_nightly_dicts) // Nbls_to_load + 1
     else:
         Nbls_to_load = len(bl_nightly_dicts) // Nblgroups
@@ -864,8 +864,10 @@ def lst_bin_files(data_files, input_cals=None, dlst=None, verbose=True, ntimes_p
                             # If uvc has Ntimes == 1, then broadcast across time will work automatically
                             uvc.select(times=uvc.time_array[tinds])
                             gains, cal_flags, _, _ = uvc.build_calcontainers()
-                        apply_cal.calibrate_in_place(data, gains, data_flags=flags, cal_flags=cal_flags,
-                                                        gain_convention=uvc.gain_convention)
+                        apply_cal.calibrate_in_place(
+                            data, gains, data_flags=flags, cal_flags=cal_flags,
+                            gain_convention=uvc.gain_convention
+                        )
                         utils.echo("Done with calibration.", verbose=verbose)
 
                     # redundantly average baselines, keying to baseline group key
@@ -973,9 +975,9 @@ def lst_bin_files(data_files, input_cals=None, dlst=None, verbose=True, ntimes_p
 
 
 def make_lst_grid(
-    dlst: float, 
-    begin_lst: float | None = None, 
-    lst_width: float = 2*np.pi, 
+    dlst: float,
+    begin_lst: float | None = None,
+    lst_width: float = 2 * np.pi,
 ) -> np.ndarray:
     """
     Make a uniform grid in local sidereal time.
@@ -986,17 +988,17 @@ def make_lst_grid(
 
     Parameters:
     -----------
-    dlst : 
+    dlst :
         The width of a single LST bin in radians. 2pi must be equally divisible
         by dlst. If not, will default to the closest dlst that satisfies this criterion that
         is also greater than the input dlst. There is a minimum allowed dlst of 6.283e-6 radians,
         or .0864 seconds.
     begin_lst
-        Beginning point for lst_grid. ``begin_lst`` must fall exactly on an LST bin 
-        given a dlst, within 0-2pi. If not, it is replaced with the closest bin. 
+        Beginning point for lst_grid. ``begin_lst`` must fall exactly on an LST bin
+        given a dlst, within 0-2pi. If not, it is replaced with the closest bin.
         Default is zero radians.
     lst_width
-        The width of the LST grid (including all bins) in radians. 
+        The width of the LST grid (including all bins) in radians.
         Default is 2pi radians. Note that regardless of this value, the final grid
         will always be equally divisible by 2pi.
 
@@ -1010,7 +1012,7 @@ def make_lst_grid(
 
     # check 2pi is equally divisible by dlst
     if not (
-        np.isclose((2 * np.pi / dlst) % 1, 0.0, atol=1e-5) 
+        np.isclose((2 * np.pi / dlst) % 1, 0.0, atol=1e-5)
         or np.isclose((2 * np.pi / dlst) % 1, 1.0, atol=1e-5)
     ):
         # generate array of appropriate dlsts
@@ -1027,7 +1029,7 @@ def make_lst_grid(
         dlst = new_dlst
 
     # make an lst grid from [0, 2pi), with the first bin having a left-edge at 0 radians.
-    lst_grid = np.arange(0, 2*np.pi - 1e-7, dlst) + dlst / 2
+    lst_grid = np.arange(0, 2 * np.pi - 1e-7, dlst) + dlst / 2
 
     # shift grid by begin_lst
     if begin_lst is not None:
@@ -1041,17 +1043,22 @@ def make_lst_grid(
         begin_lst = 0.0
 
     lst_grid = lst_grid[lst_grid < (begin_lst + lst_width)]
-    
+
     return lst_grid
 
 
-def sigma_clip(array: np.ndarray | np.ma.MaskedArray, sigma: float=4.0, min_N: int=4, axis: int = 0):
+def sigma_clip(
+    array: np.ndarray | np.ma.MaskedArray,
+    sigma: float = 4.0,
+    min_N: int = 4,
+    axis: int = 0
+):
     """
     One-iteration robust sigma clipping algorithm.
 
     Parameters
     ----------
-    array 
+    array
         ndarray of *real* data.
     sigma
         sigma threshold to cut above.
@@ -1063,8 +1070,8 @@ def sigma_clip(array: np.ndarray | np.ma.MaskedArray, sigma: float=4.0, min_N: i
 
     Output
     ------
-    clip_flags 
-        A boolean array with same shape as input array, 
+    clip_flags
+        A boolean array with same shape as input array,
         with clipped values set to True.
     """
     # ensure array is an array
@@ -1092,11 +1099,11 @@ def sigma_clip(array: np.ndarray | np.ma.MaskedArray, sigma: float=4.0, min_N: i
 
 
 def gen_bl_nightly_dicts(
-    hds: list[io.HERAData], 
-    bl_error_tol: float=1.0, 
-    include_autos: bool=True, 
-    redundant: bool=False, 
-    ex_ant_yaml_files: list[str] | None=None
+    hds: list[io.HERAData],
+    bl_error_tol: float = 1.0,
+    include_autos: bool = True,
+    redundant: bool = False,
+    ex_ant_yaml_files: list[str] | None = None
 ) -> list[dict[int, list[tuple[int, int]]]]:
     """
     Helper function to generate baseline dicts to keep track of reds between nights.
@@ -1104,7 +1111,7 @@ def gen_bl_nightly_dicts(
     Parameters
     ----------
     hds
-        A list of HERAData objects. Can have no data loaded (preferable) and should 
+        A list of HERAData objects. Can have no data loaded (preferable) and should
         refer to single files.
     bl_error_tol
         Baselines whose vector difference are within this tolerance are considered
@@ -1112,8 +1119,8 @@ def gen_bl_nightly_dicts(
     include_autos
         Whether to include autos in bl_nightly_dicts.
     redundant
-        If True, ``bl_nightly_dict`` stores redundant group. If False, each 
-        ``bl_nightly_dict`` will contain a length-1 group for each baseline over all 
+        If True, ``bl_nightly_dict`` stores redundant group. If False, each
+        ``bl_nightly_dict`` will contain a length-1 group for each baseline over all
         the nights.
     ex_ant_yaml_files
         A list of paths to yaml files with antennas to throw out on each night.
@@ -1122,7 +1129,7 @@ def gen_bl_nightly_dicts(
     Outputs:
     --------
     If redundant:
-        list of dictionaries of the form 
+        list of dictionaries of the form
         ``{0: [(a0, b0), (a0, c0)...], 1: [(a1, b1),.., ], ... Nnight: [(ANnight, BNnight), ...,]}.
         Each dictionary represents a unique baseline length and orientation.
         where the key of each dictionary is an index for each night to be LST binned and each value
